feat : Calculator 모든 기능 구현

+++ Calculator/src/main/java/org/example/io/ConsoleOutput.java
@@ -9,7 +9,7 @@ public class ConsoleOutput implements Output{
         System.out.println();
     }
     @Override
-    public void printCalculatedResult(int result) {
+    public void printCalculatedResult(long result) {
         System.out.println(result);
         System.out.println();
     }
