+++ Calculator/src/main/java/org/example/io/Output.java
@@ -4,7 +4,7 @@ import org.example.repository.Repository;
 
 public interface Output {
     void printMenu(String menu);
-    void printCalculatedResult(int result);
+    void printCalculatedResult(long result);
     void printSavedResults(Repository repository);
