+++ Calculator/src/main/java/org/example/repository/Repository.java
@@ -1,7 +1,7 @@
 package org.example.repository;
 
 public interface Repository {
-    void saveResult(String input, int output);
+    void saveResult(String input, long output);
 
     void showSavedResults();
 }
